% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotEnsemble.R
\name{ensemble_plot_index}
\alias{ensemble_plot_index}
\title{Plot Indices}
\usage{
ensemble_plot_index(summaryoutput, varlist, indexfleets = 1, verbose = TRUE)
}
\arguments{
\item{summaryoutput}{summaryoutput}

\item{varlist}{variable list}

\item{indexfleets}{Fleet vector index}

\item{verbose}{Option to output messages to Rconsole}
}
\description{
function to plot different fits to a single index of abundance
}
\keyword{internal}
\keyword{ssplot}
