% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss3_exe.R
\name{get_ss3_exe}
\alias{get_ss3_exe}
\alias{ss3_exe_windows}
\alias{ss3_exe_darwin}
\alias{ss3_exe_linux}
\title{Download the Stock Synthesis (SS3) executable}
\usage{
get_ss3_exe(dir = NULL, version = NULL)

ss3_exe_windows(dir, tag)

ss3_exe_darwin(dir, tag, arch = c("aarch64", "x86_64"))

ss3_exe_linux(dir, tag)
}
\arguments{
\item{dir}{Target Directory}

\item{version}{A character string of the executable version tag to download
(e.g.'v3.30.20' or 'v3.30.18'). A list of tags is available at
https://github.com/nmfs-ost/ss3-source-code/tags}

\item{tag}{Target ss3 Version release}

\item{arch}{arch}
}
\value{
A string of the file path to the downloaded executable

A string of the file path to the downloaded executable

A string of the file path to the downloaded executable

A string of the file path to the downloaded executable
}
\description{
The \code{get_ss3_exe()} function uses the \code{gh} package to get either
the latest release (if version = NULL) or the specified version of the Stock
Synthesis executable for the appropriate operating system to the directory \code{dir}
(if dir = NULL, then the executable is downloaded to the working directory).
To view the version tags available go to
https://github.com/nmfs-ost/ss3-source-code/tags
}
\details{
Downloads the SS3 executable according to specified version and the user
operating system.
}
\examples{
\dontrun{
get_ss3_exe()
get_ss3_exe(version = "v3.30.18")
}
}
\author{
Elizabeth F. Gugliotti
}
\keyword{internal}
