% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmcSimple}
\alias{mcmcSimple}
\title{MCMC output for simulated Simple Stock Synthesis model.}
\format{
Large dataframe of MCMC parameter estimates with nrow = number of values and ncol = number of parameters
\describe{
\item{mcmcSimple}{mcmcSimple}
}
}
\source{
MCMC evaluated with Stock Synthesis
}
\usage{
mcmcSimple
}
\description{
MCMC output read by \code{\link[r4ss]{SSgetMCMC}()}
}
\author{
R. Methot,
}
\keyword{datasets}
