% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pac.hke}
\alias{pac.hke}
\title{2020 Pacific Hake Stock Synthesis model}
\format{
\code{r4ss} list objects
\describe{
 \item{ss3phke}{ss3 rep file read with r4ss::SS_output()}
 \item{retro.phk}{list of retrospective rep files read with r4ss::SSgetoutput()}
 \item{aspm.phk}{list of referece and aspm rep files created with r4ss::SSsummarize()}
}
}
\source{
Joint Technical Committee of the U.S. and Canada Pacific Hake/Whiting 
Agreement, National Marine Fisheries Service and Fisheries and Oceans Canada
}
\usage{
pac.hke
}
\description{
Stock Synthesis (V3.30) base-case run, list of retrospective runs,
 list of base-case and aspm runs
}
\author{
Grandin, C.J., K.F. Johnson, A.M. Edwards, and A.M. Berger (2020)
}
