% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRunstest.R
\name{plot_runstest}
\alias{plot_runstest}
\title{plot function for runs test plot}
\usage{
plot_runstest(
  resdat,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{resdat}{residual data.frame}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
Residual diagnostics with runs test p-value and 3xsigma limits for Indices and meanL
}
