% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rc}
\alias{rc}
\title{r4ss color palette}
\usage{
rc(n, alpha = 1)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transluscency}
}
\value{
vector of color codes
}
\description{
r4ss color palette
}
\keyword{ssplot}
\keyword{utils}
