% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retro.phk}
\alias{retro.phk}
\title{2020 Pacific Hake Retrospective Runs}
\format{
List of 8 retrospective runs generated using \pkg{r4ss}:

\describe{
 \item{replist1}{replist1}
 \item{replist2}{replist2}
 \item{replist3}{replist3}
 \item{replist4}{replist4}
 \item{replist5}{replist5}
 \item{replist6}{replist6}
 \item{replist7}{replist7}
 \item{replist8}{replist8}
}
}
\source{
2020 Pacific hake/whiting Stock Synthesis assessment. Joint Technical 
Committee of the U.S. and Canada Pacific Hake/Whiting Agreement, 
National Marine Fisheries Service and Fisheries and Oceans Canada
}
\usage{
retro.phk
}
\description{
List of retrospective runs with \code{\link[r4ss]{SS_doRetro}()} as read by \code{\link[r4ss]{SSgetoutput}()}
}
\author{
Grandin, C.J., K.F. Johnson, A.M. Edwards, and A.M. Berger. (2020). 
Status of the Pacific Hake (whiting) stock in U.S. and Canadian waters in 2020. 
Prepared by the Joint Technical Committee of the U.S. and Canada 
Pacific Hake/Whiting Agreement, National Marine Fisheries Service and 
Fisheries and Oceans Canada. 273 p.
}
\keyword{datasets}
