% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retro.sma}
\alias{retro.sma}
\title{ICCAT 2017 North Atlantic Shortfin Mako retrospective runs}
\format{
Large List of 6 elements generated using \pkg{r4ss}
\describe{
 \item{replist1}{replist1}
 \item{replist2}{replist2}
 \item{replist3}{replist3}
 \item{replist4}{replist4}
 \item{replist5}{replist5}
 \item{replist6}{replist6}
}
}
\source{
ICCAT 2017. Report of the 2017 Shortfin Mako Assessment Meeting (Madrid, Spain 12–16 June 2017).Collect. Vol. Sci. Pap.
}
\usage{
retro.sma
}
\description{
A list of retrospective runs with \code{\link[r4ss]{SS_doRetro}()} as read by 
\code{\link[r4ss]{SSgetoutput}()}
}
\author{
Courtney, C., D., Carvalho, C. Winker, H., and LT Kell. (2020). Examples of Stock Synthesis diagnostic methods and results implemented for previously completed North Atlantic shortfin mako Stock Synthesis model runs. Collect. Vol. Sci. Pap. ICCAT, 76(10): 173-234.
}
\keyword{datasets}
