% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss3her}
\alias{ss3her}
\title{GOB Herring Synthesis model output}
\format{
List of 188
}
\usage{
ss3her
}
\description{
Stock Synthesis output read by \code{\link[r4ss]{SS_output}()}
}
\keyword{datasets}
