% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SScompsTA1.8.R
\name{SScompsTA1.8}
\alias{SScompsTA1.8}
\title{TA1.8 Francis function modified from r4ss::SSMethod.TA1.8 to create mean length/age}
\usage{
SScompsTA1.8(
  ss3rep,
  type = c("len", "age", "size", "con")[1],
  fleet = NULL,
  plotit = FALSE,
  maxpanel = 1000
)
}
\arguments{
\item{ss3rep}{Stock Synthesis output as read by r4SS function SS_output}

\item{type}{either 'len' (for length composition data), 'size' (for
generalized size composition data), 'age' (for age composition data),
or 'con' (for conditional age at length data)}

\item{fleet}{vector of one or more fleet numbers whose data are to
be analysed simultaneously (the output N multiplier applies
to all fleets combined)}

\item{plotit}{if TRUE, make an illustrative plot like one or more
panels of Fig. 4 in Francis (2011).}

\item{maxpanel}{maxpanel. Default 1000}
}
\value{
data.frame of observed, predicted mean length/age and residuals
}
\description{
Uses method TA1.8 (described in Appendix A of Francis 2011) to do
stage-2 weighting of composition data from a Stock Synthesis model.
The main purpose is to create obs, exp and residuals of mean L/age 
Please see r4ss::SSMethod.TA1.8 for full set of intended options
}
\references{
Francis, R.I.C.C. (2011). Data weighting in statistical
fisheries stock assessment models. Canadian Journal of
Fisheries and Aquatic Sciences 68: 1124-1138.
}
\author{
Chris Francis, Andre Punt, Ian Taylor (modified by Henning Winker)
}
