% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdeltaMVLN_test.R.R
\name{SSdeltaMVLN_test}
\alias{SSdeltaMVLN_test}
\title{SSdeltaMVLN_test()}
\usage{
SSdeltaMVLN_test(
  ss3rep,
  Fref = NULL,
  years = NULL,
  mc = 5000,
  weight = 1,
  run = "MVLN",
  plot = TRUE,
  addprj = FALSE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE
)
}
\arguments{
\item{ss3rep}{from r4ss::SSgetoutput()$replist1}

\item{Fref}{Choice of Fratio c("MSY","Btgt), correponding to F_MSY and F_Btgt}

\item{years}{single year or vector of years for mvln}

\item{mc}{number of monte-carlo simulations}

\item{weight}{weighting option for model ensembles weight*mc}

\item{run}{qualifier for model run}

\item{plot}{option to show plot}

\item{addprj}{include forecast years}

\item{ymax}{ylim maximum}

\item{xmax}{xlim maximum}

\item{verbose}{Report progress to R GUI?}

\item{legendcex=1}{Allows to adjust legend cex}
}
\value{
output list of kobe objects and mle's
}
\description{
function to generatbe kobe pdfs from a Multivariate Log-Normal Distribution
including plotting option
}
\author{
Henning Winker (JRC-EC)
}
