% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdiagsMCMC.R
\name{SSdiagsMCMC}
\alias{SSdiagsMCMC}
\title{SSdiagsMCMC()}
\usage{
SSdiagsMCMC(
  mcmc,
  ss3rep,
  Fref = NULL,
  years = NULL,
  run = "MCMC",
  thin = 1,
  plot = TRUE,
  addprj = FALSE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE
)
}
\arguments{
\item{mcmc}{file path for folder with the derived_posteriors.sso file}

\item{ss3rep}{from r4ss::SS_output}

\item{Fref}{Choice of Fratio c("MSY","Btgt), correponding to F_MSY and F_Btgt}

\item{years}{single year or vector of years for mvln}

\item{run}{qualifier for model run}

\item{thin}{option to use additional thinning}

\item{plot}{option to show plot}

\item{addprj}{include forecast years}

\item{ymax}{ylim maximum}

\item{xmax}{xlim maximum}

\item{legendcex}{Allows to adjust legend cex}

\item{verbose}{Report progress to R GUI?}
}
\value{
output list of quant mcmc posteriors and mle's
}
\description{
function to read mcmc file outputs for Kobe and SSplotEnsemble() plotting
}
\author{
Henning Winker (JRC-EC), Massimiliano and Laurence Kell (Sea++)
}
\keyword{MCMC}
\keyword{diags}
