% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdiagsMCMC_ancient.R
\name{SSdiagsMCMC_ancient}
\alias{SSdiagsMCMC_ancient}
\title{MCMC Joint Distributions}
\usage{
SSdiagsMCMC_ancient(
  mcmc,
  ss3rep,
  Fref = NULL,
  years = NULL,
  run = "MCMC",
  thin = 1,
  plot = TRUE,
  addprj = FALSE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE
)
}
\arguments{
\item{mcmc}{file path for folder with the derived_posteriors.sso file}

\item{ss3rep}{from r4ss::SS_output}

\item{Fref}{Choice of Fratio c("MSY","Btgt), correponding to F_MSY and F_Btgt}

\item{years}{single year or vector of years for mvln}

\item{run}{qualifier for model run}

\item{thin}{option to use additional thinning}

\item{plot}{option to show plot}

\item{addprj}{include forecast years}

\item{ymax}{ylim maximum}

\item{xmax}{xlim maximum}

\item{legendcex}{Allows to adjust legend cex}

\item{verbose}{Report progress to R GUI?}
}
\value{
output list maximum likelihood estimates and the MCMC posterier distributions of the Kobe values and a Kobe plot
}
\description{
A function to generate joint distributions for stock status ratios (SSB/SSBref and F/Fref where ref can be MSY, SSB40, F40, etc.) using MCMC. The function produces a Kobe plot, maximum likelihood estimates and the MVLN Monte-Carlo distributions of the Kobe values which can be input into SSplotEnsemble(). Use for Stock Synthesis models older than 3.24, for later versions use SSdiagsMCMC().
}
\author{
Henning Winker (JRC-EC), Massimiliano and Laurence Kell (Sea++)
}
