% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdiagsMCMC_old.R
\name{SSdiagsMCMC_old}
\alias{SSdiagsMCMC_old}
\title{SSdiagsMCMC_old()}
\usage{
SSdiagsMCMC_old(
  mcmcdir,
  Bref = c("MSY", "Btrg"),
  Fref = c("MSY", "Ftrg"),
  run = "MCMC",
  Fstarter = c("_abs_F", "(F)/(Fmsy)", "Fstd_Btgt"),
  forecast = FALSE,
  plot = TRUE,
  thin = 1,
  biomass = "SSB",
  refs = c("SSB_unfished", "SSB_MSY", "SSB_Btgt", "SSB_SPR", "SPR_MSY", "Fstd_MSY",
    "Fstd_SPR", "Fstd_Btgt", "Recr_unfished", "B_MSY.SSB_unfished", "Dead_Catch_MSY",
    "Ret_Catch_MSY")
)
}
\arguments{
\item{mcmcdir}{file path for folder with the derived_posteriors.sso file}

\item{Bref}{Choice of reference point for stock SSB/X c("MSY","Btrg")}

\item{Fref}{Choice of reference point for stock SSB/XFref=c("MSY","Ftrg")}

\item{run}{qualifier for model run}

\item{Fstarter}{starter settings for  c("_abs_F","(F)/(Fmsy)","Fstd_Btgt") see SSsettingsBratioF()}

\item{forecast}{option to include forecasts TRUE/FALSE}

\item{thin}{option to use additional thinning}

\item{biomass}{the function is only tested to run with default biomass = "SSB"}

\item{refs}{required reference quantaties}

\item{Plot}{option to plot results with SSplotEnsemble()}
}
\value{
list of (1) all quataties "sims" (2) object of SSplotEnsemble() "kb" and settings
}
\description{
function to read mcmc file outputs for Kobe and SSplotEnsemble() plotting
}
\author{
Henning Winker (JRC-EC), Massimiliano and Laurence Kell (Sea++)
}
