% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SSdiagsStep2Year}
\alias{SSdiagsStep2Year}
\title{SSdiagsStep2Year()}
\usage{
SSdiagsStep2Year(ss3out, time.steps = 0.25, styr, season = 1)
}
\arguments{
\item{ss3out}{outputs from r4ss::SS_output(), r4ss::SSsummarize() or ss3diags::SSretroComps()}

\item{time.steps}{time steps behind yrs e.g. 0.25 for quarterly}

\item{styr}{start year step (e.g. 1950)}

\item{season}{reference Season for retrospective}
}
\value{
Reformatted Rep file outputs
}
\description{
Function to convert non-annual into annual time-steps for retros and cpue residuals
}
