% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSgetComps.R
\name{SSgetComps}
\alias{SSgetComps}
\title{SSgetComps}
\usage{
SSgetComps(dirvec = getwd(), ncols = 400, covar = FALSE, verbose = FALSE)
}
\arguments{
\item{dirvec}{A vector of directories (full path or relative to working}

\item{ncols}{Maximum number of columns in Report.sso (same input as for}

\item{covar}{Choice to read or not read covar.sso output (saves time and}

\item{verbose}{Print various messages to the command line as the function}
}
\value{
list of observed and expected mean Length/age comps (c.f. ss3rep$cpue)
}
\description{
Get observed and expected mean length/age from multiple Stock Synthesis models.
Wrapper that uses r4ss::SS_output() and SScompsTA1.8()
}
\author{
Henning Winker
}
