% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShcbias.R
\name{SShcbias}
\alias{SShcbias}
\title{SShcbias()}
\usage{
SShcbias(
  summaryoutput,
  quants = c("SSB", "F"),
  models = "all",
  endyrvec = "default",
  verbose = TRUE
)
}
\arguments{
\item{summaryoutput}{List created by r4ss::SSsummarize()}

\item{quants}{SSB or F quantaty}

\item{models}{Optional subset of the models described in
r4ss function summaryoutput().  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{verbose}{Report progress to R GUI?}
}
\description{
Computes Mohn's Rho and one step ahead forecast bias (forecastrho)
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
