% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmase.R
\name{SSmase}
\alias{SSmase}
\title{SSmase() computes MASE for one-step ahead hindcasting cross-validations of indices}
\usage{
SSmase(
  retroSummary,
  subplots = c("cpue", "len", "age"),
  Season = "default",
  models = "all",
  endyrvec = "default",
  indexselect = NULL,
  verbose = TRUE
)
}
\arguments{
\item{retroSummary}{List created by r4ss::SSsummarize()}

\item{Season}{option to specify Season - Default uses first available, i.e. usual Seas = 1}

\item{models}{Optional subset of the models described in
r4ss function summaryoutput().  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{indexselect}{= Vector of fleet numbers for each model for which to compare}

\item{verbose}{Report progress to R GUI?}

\item{indexfleets}{CHECK IF NEEDED or how to adjust indexfleets}
}
\description{
MASE for one-step ahead hindcasting cross-validations and computes MASE from prediction redisuals
}
\author{
Henning Winker (JRC-EC) and Laurence Kell (Sea++)
}
