% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplot_eqcurves.R
\name{SSplot_eqcurves}
\alias{SSplot_eqcurves}
\title{SSplot_eqcurves()}
\usage{
SSplot_eqcurves(
  ss3rep,
  Fref = c("MSY", "Btgt", "SPR", "F01")[2],
  msyline = NULL,
  cex = 0.7
)
}
\arguments{
\item{ss3rep}{from r4ss::SS_output}

\item{Fref}{Choice of Fratio c("MSY","Btgt","SPR","F01"), correponding to F_MSY and F_Btgt}
}
\value{
base plot
}
\description{
function to plot equilibrium yield curves
}
\examples{
# MSY refpoints with pretty good yield 80\% line
SSplot_eqcurves(ss3sma,Fref="MSY",msyline=0.8) 
# SSB40 refpoints 
SSplot_eqcurves(ss3phk,Fref="Btgt") 
# SPR40 refpoints 
SSplot_eqcurves(ss3phk,Fref="SPR") 
}
\author{
Henning Winker (GFCM)
}
