% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotEnsemble.R
\name{SSplotEnsemble}
\alias{SSplotEnsemble}
\title{Multivariate lognormal (MVLN) Monte-Carlo distribution ensemble plot}
\usage{
SSplotEnsemble(
  kb,
  summaryoutput,
  subplots = c("stock", "harvest", "SSB", "F", "Recr", "Catch"),
  models = "all",
  quantiles = c(0.025, 0.975),
  ylabs = NULL,
  endyrvec = "default",
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = FALSE,
  use_pdf = FALSE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = NULL,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "l",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = NULL,
  shadealpha = 0.3,
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  indexQlabel = TRUE,
  indexQdigits = 4,
  legendindex = NULL
)
}
\arguments{
\item{kb}{SSdeltaMVLN $kb type output}

\item{summaryoutput}{List created by r4ss::SSummarize(). TODO: Verify}

\item{subplots}{option to "Bratio","Fvalue","SSB", "F", "Recr","Catch"}

\item{models}{option to manually subset the models in kb$run}

\item{quantiles}{quantile TODO TODO. Default is (.025,.075)}

\item{ylabs}{yaxis labels for quants
final year of values to show for each model. By default it is set to the}

\item{endyrvec}{ending year specified in each model.}

\item{plot}{Option to draw subplots and plot in the interface. Deprecated. Option to disable will be removed in future version.}

\item{print}{print to PNG files? Deprecated. Please use print_plot.}

\item{print_plot}{print to PNG files?}

\item{png}{output in PNG format. Deprecated. Please use use_png.}

\item{use_png}{Draw plots in PNG format}

\item{pdf}{not tested for TRUE. Deprecated. Please use use_pdf.}

\item{use_pdf}{option for pdf plots (not tested for TRUE)}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis mark at final
year in timeseries plots.}

\item{xlim}{Range of years. Optional parameter. NULL by default.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space}

\item{xaxs}{Choice of xaxs parameter (see ?par for more info)}

\item{yaxs}{Choice of yaxs parameter (see ?par for more info)}

\item{xylabs}{xylabs X asis and Y axis labels TODO TODO. Default in NULL !}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}

\item{uncertainty}{uncertainty TODO TODO. Default is TRUE}

\item{legend}{Add a legend to plot. TRUE by default.}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector
of two numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See ?legend for more info on the string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have
the legend display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendcex}{Allows to adjust legend cex}

\item{legendsp}{Space between legend labels}

\item{pwidth}{Width of plot}

\item{pheight}{Height of plot}

\item{punits}{Units for PNG file}

\item{res}{Resolution for PNG file}

\item{ptsize}{Point size for PNG file}

\item{cex.main}{Character expansion for plot titles}

\item{plotdir}{Directory where PNG or PDF files will be written. By default
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{par}{list of graphics parameter values passed to par() function}

\item{verbose}{Report progress to R GUI?}

\item{shadecol}{uncertainty shading of hcxval horizon}

\item{shadealpha}{Transparency adjustment used to make default shadecol}

\item{new}{Create new empty plot window. Deprecated.}

\item{add}{surpresses par() to create multiplot figs}

\item{mcmcVec}{mcmc vector TODO TODO. Default is FALSE}

\item{indexQlabel}{indexQlabel TODO TODO. Default is TRUE}

\item{indexQdigits}{indexQdigits TODO TODO. Default is 4}

\item{legendindex}{Allows to add lengend for selected indices (plots)}
}
\description{
Plots the Distributions for SSB/SSB40, F/FSPR40, SSB, F, Recruitment and Catch trajectories for the target model.
}
\author{
Mostly adopted from r4ss::SSplotComparisons by Taylor et al
}
\keyword{ssplot}
