% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotGeneric.R
\name{SSplotGenericPar}
\alias{SSplotGenericPar}
\title{Graphical Parameters for Stock Synthesis Generic plots.}
\usage{
SSplotGenericPar(
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = NULL,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "l"
)
}
\arguments{
\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{\link[base:logical]{Logical} flag: set TRUE or FALSE to switch to turn
on/off extra axis mark at final year in timeseries plots.}

\item{xlim}{Optional, years to use for x-axis. Default value NULL (or "default"),
uses all years available.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space.}

\item{xaxs}{Choice of xaxs parameter See \code{?par} for more info.}

\item{yaxs}{Choice of yaxs parameter. See \code{?par} for more info.}

\item{xylabs}{\link[base:logical]{Logical} flag: set TRUE or FALSE to include x- and
y-axis labels to the plot. Defaults to TRUE}

\item{type}{The type of plot to be drawn. For more details, see \code{\link[base:plot]{plot}}.}
}
\description{
Generalized version of ss3diags SSplots functions to specify to \code{par()} or
to pass plot attributes. Not intended for production.
}
