% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotGeneric.R
\name{SSplotGenericUncertainty}
\alias{SSplotGenericUncertainty}
\title{Uncertainty parameters for Stock Synthesis Generic plots.}
\usage{
SSplotGenericUncertainty(
  uncertainty = TRUE,
  mcmcVec = FALSE,
  endyrvec = "default",
  models = "all"
)
}
\arguments{
\item{uncertainty}{Logical flag to enable plots with uncertainty intervals. Either a single
TRUE/FALSE value, or a vector of TRUE/FALSE values around SSB or F
estimated timeseries, or a set of integers corresponding to
the choice of models.}

\item{mcmcVec}{Logical vector of TRUE/FALSE values (or single value) indicating
whether input values are from MCMC or to use normal distribution around
MLE.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By \code{"default"} it is set to the
ending year specified in each model.}

\item{models}{Optional subset of the models of \code{summaryoutput} (or a similar
field with a different name): a list created by the function
\link[r4ss:SSsummarize]{r4ss::SSsummarize}. Either \code{"all"} or a vector of numbers indicating
columns in summary tables.}
}
\description{
Generalized version of ss3diags SSplots functions to specify uncertainty.
Not intended for production.
}
\keyword{internal}
