% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotJABBAres.R
\name{SSplotJABBAres}
\alias{SSplotJABBAres}
\title{SSplotJABBAres()}
\usage{
SSplotJABBAres(
  ss3rep = ss3diags::ss3sma,
  subplots = c("cpue", "len", "age")[1],
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  indexselect = NULL,
  miny = 3,
  col = NULL,
  pch = 21,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  xlim = "default",
  ylimAdj = 1.1,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  legend = TRUE,
  legendlabels = "default",
  legendloc = "bottomleft",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  boxcol = grey(0.8, 0.5),
  new = TRUE,
  add = FALSE
)
}
\arguments{
\item{ss3rep}{from r4ss::SS_output}

\item{subplots}{optional use of cpue and comp data (only tested for length)}

\item{plot}{Deprecated. Plots (and subplots) are drawn to the active plot device 
by default (TRUE), and the option to disable this, via FALSE, is unused.}

\item{print}{Deprecated. Please use 'print_plot'.}

\item{print_plot}{Option to print to PNG files}

\item{png}{Deprecated. please use 'use_png'.}

\item{use_png}{Draw plots in PNG format}

\item{pdf}{Deprecated. Please use 'use_pdf'.}

\item{use_pdf}{option for pdf plots (not tested for TRUE)}

\item{indexselect}{Vector of fleet numbers for each model for which to compare}

\item{miny}{minimum abs values of ylim}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis mark at final
year in timeseries plots.}

\item{xlim}{xlim TODO TODO,}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space}

\item{xaxs}{Choice of xaxs parameter (see ?par for more info)}

\item{yaxs}{Choice of yaxs parameter (see ?par for more info)}

\item{xylabs}{xylabs TODO TODO. Default is TRUE}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}

\item{legend}{Option to add a legend. TRUE by default.}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector
of two numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See ?legend for more info on the string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have
the legend display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendcex}{Allows to adjust legend cex}

\item{legendsp}{Space between legend labels}

\item{legendindex}{Allows to add lengend for selected indices (plots)}

\item{pwidth}{Width of plot}

\item{pheight}{Height of plot}

\item{punits}{Units for PNG file}

\item{res}{Resolution for PNG file}

\item{ptsize}{Point size for PNG file}

\item{cex.main}{Character expansion for plot titles}

\item{plotdir}{Directory where PNG or PDF files will be written. By default
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{par}{list of graphics parameter values passed to par() function}

\item{verbose}{Report progress to R GUI?}

\item{boxcol}{color boxes}

\item{new}{Deprecated. New plot windows are created by default (TRUE), and the 
option to disable this, via FALSE, is unused.}

\item{add}{surpresses par() to create multiplot figs}
}
\description{
plots residuals for all indices as boxplot with a loess showing systematic trends
}
\author{
Henning Winker (JRC-EC)
}
\keyword{ssplot}
