% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSkobe.R
\name{SSplotKobe}
\alias{SSplotKobe}
\title{KOBE phase plot}
\usage{
SSplotKobe(
  kb,
  joint = TRUE,
  year = NULL,
  posterior = c("points", "kernel"),
  ci.levels = c(0.5, 0.8, 0.95),
  path = TRUE,
  xlab = expression(SSB/SSB[MSY]),
  ylab = expression(F/F[MSY]),
  pt.cex = 0.8,
  ylim = NULL,
  xlim = NULL,
  fill = TRUE,
  legend = TRUE,
  legendpos = "right",
  legendcex = 0.7,
  legendruns = TRUE,
  yr.label = TRUE,
  yr.int = 5,
  verbose = TRUE
)
}
\arguments{
\item{kb}{output from SSdeltaMVLN()$kb}

\item{joint}{option FALSE shows individual runs}

\item{year}{option to choose year for kobe, last year is default}

\item{posterior}{visualization of posterior c("points","kernel")}

\item{ci.levels}{option for kernel with default ci.levels=c(0.50,0.80,0.95)}

\item{path}{shows historical years}

\item{xlab}{graphic parameter}

\item{ylab}{graphic parameter}

\item{pt.cex}{option change relative size of points (default 0.8)}

\item{ylim}{graphic parameter}

\item{xlim}{graphic parameter}

\item{fill}{shows color-coded quadrants}

\item{legend}{graphic parameter}

\item{legendpos}{legend position}

\item{legendcex}{lengend size}

\item{legendruns}{show legend for run lables}

\item{yr.label}{show year along trajectory}

\item{yr.int}{= year interval along trajectory}

\item{verbose}{}
}
\value{
Kobe Quadrant percentages
}
\description{
plots the stock status uncertaint over SSB/SSBmsy and F/Fmsy
}
