% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRetro.R
\name{SSplotRetro}
\alias{SSplotRetro}
\title{SSplotRetro() Retrospective-Forecast with one-step ahead hindcasting}
\usage{
SSplotRetro(
  summaryoutput,
  subplots = c("SSB", "F"),
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  models = "all",
  endyrvec = "default",
  xlim = NULL,
  xmin = NULL,
  labels = NULL,
  ylim = NULL,
  forecast = TRUE,
  forecastrho = TRUE,
  showrho = TRUE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.7,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = grey(0.4, 0.6),
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  shadecol1 = grey(0.5, 0.4),
  indexQlabel = TRUE,
  indexQdigits = 4,
  shadealpha = 0.3
)
}
\arguments{
\item{summaryoutput}{List created by r4ss::SSsummarize()}

\item{subplots}{Vector of subplots to be created}

\item{plot}{Option to draw subplots and plot in the interface. Deprecated. Option to disable will be removed in future version.}

\item{print}{print to PNG files? Deprecated. Please use print_plot.}

\item{print_plot}{Option to print to PNG files}

\item{png}{png plots. Deprecated, please use use_png}

\item{use_png}{Draw plots in PNG format}

\item{pdf}{PDF plots (not tested for TRUE). Deprecated. Please use use_pdf.}

\item{use_pdf}{option for pdf plots (not tested for TRUE)}

\item{models}{Optional subset of the models described in
r4ss function summaryoutput().  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{xlim}{optional xlim, which overwrites xmin}

\item{xmin}{optional minimum year shown in plot (default first yr)}

\item{labels}{yaxis lable for biomass (bony fish and sharks)}

\item{ylim}{option to specify ylim range}

\item{forecast}{if true one-step ahead forecasts are shown in plot}

\item{forecastrho}{if true one-step ahead forecast rho value is denoted in plot}

\item{showrho}{showrho TODO TODO. Defaults to TRUE}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis mark at final
year in timeseries plots.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space}

\item{xaxs}{Choice of xaxs parameter (see ?par for more info)}

\item{yaxs}{Choice of yaxs parameter (see ?par for more info)}

\item{xylabs}{Draw x and y axis lables (?) TODO TODO. Defaults to TRUE}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}

\item{uncertainty}{uncertainty TODO TODO. Defaults to TRUE.}

\item{legend}{Add a legend?}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector
of two numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See ?legend for more info on the string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have
the legend display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendcex}{Allows to adjust legend cex}

\item{legendsp}{Space between legend labels}

\item{legendindex}{Allows to add lengend for selected indices (plots)}

\item{pwidth}{Width of plot}

\item{pheight}{Height of plot}

\item{punits}{Units for PNG file}

\item{res}{Resolution for PNG file}

\item{ptsize}{Point size for PNG file}

\item{cex.main}{Character expansion for plot titles}

\item{plotdir}{Directory where PNG or PDF files will be written. By default
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{par}{list of graphics parameter values passed to par() function}

\item{verbose}{Report progress to R GUI?}

\item{shadecol}{uncertainty shading of hcxval horizon}

\item{new}{Create new empty plot window. Deprecated.}

\item{add}{surpresses par() to create multiplot figs}

\item{mcmcVec}{NOT TESTED Vector of TRUE/FALSE values (or single value) indicating}

\item{shadecol1}{uncertainty shading of early years not affected by hindcast}

\item{indexQlabel}{Add catchability to legend in plot of index fits (TRUE/FALSE)?}

\item{indexQdigits}{Number of significant digits for catchability in legend}

\item{shadealpha}{shadealpha TODO TODO. Defalut to 0.3}
}
\description{
Plots retrospective pattern, including (optional) one-step ahead forecast and computes Mohn's Rho
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
\keyword{retro}
\keyword{ssplot}
