% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRunstest.R
\name{SSplotRunstest}
\alias{SSplotRunstest}
\title{Residual Diagnostics}
\usage{
SSplotRunstest(
  ss3rep = ss3diags::ss3sma,
  mixing = "less",
  subplots = c("cpue", "len", "age", "con")[1],
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  indexselect = NULL,
  miny = 1,
  pch = 21,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = "default",
  ylimAdj = 1.4,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  legend = TRUE,
  legendloc = "top",
  legendcex = 1,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  new = TRUE,
  add = FALSE
)
}
\arguments{
\item{ss3rep}{Stock Synthesis output as read by r4SS function SS_output}

\item{mixing}{c("less","greater","two.sided"). Default less is checking for positive autocorrelation only}

\item{subplots}{optional 'cpue' for index data, 'len' for length composition data, 'size' for
generalized size composition data, 'age' for age composition data,
or 'con' for conditional age at length data}

\item{plot}{plot to active plot device?}

\item{print}{print to PNG files? Deprecated. Please use print_plot.}

\item{print_plot}{Option to print to PNG files}

\item{png}{png plots. Deprecated, please use use_png}

\item{use_png}{Draw plots in PNG format}

\item{pdf}{PDF plots. Deprecated. Please use use_pdf.}

\item{use_pdf}{option for pdf plots}

\item{indexselect}{Vector of fleet numbers for each model for which to compare}

\item{miny}{minimum abs values of ylim}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis mark at final
year in timeseries plots.}

\item{xlim}{Optional, values for x-axis range of years to display on plot. Default = "default" displays all years of available data.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space}

\item{xaxs}{Choice of xaxs parameter (see ?par for more info)}

\item{yaxs}{Choice of yaxs parameter (see ?par for more info)}

\item{xylabs}{TRUE or FALSE, include x- and y-axis labels}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}

\item{legend}{Add a legend?}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector}

\item{legendcex}{Allows to adjust legend cex}

\item{pwidth}{Width of plot}

\item{pheight}{Height of plot}

\item{punits}{Units for PNG file}

\item{res}{Resolution for PNG file}

\item{ptsize}{Point size for PNG file}

\item{cex.main}{Character expansion for plot titles}

\item{plotdir}{Directory where PNG or PDF files will be written. By default
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{par}{list of graphics parameter values passed to par() function}

\item{verbose}{TRUE or FALSE, should the progress be reported to R GUI?}

\item{new}{Create new empty plot window (TRUE or FALSE)}

\item{add}{suppresses par() to create multiplot figs}
}
\value{
a dataframe with runs test p-value, if the test has passed or failed, 3x sigma high and low limits, and the type of data used. Rows are for each fleet. Note, runs test passed if p-value > 0.05 (residuals are random) and failed if p-value < 0.5 (residuals are not random)
}
\description{
Function for residual diagnostics. Plots residuals and 3x sigma limits for indices or mean age or length and outputs a runs test table. Note, if you do not want to plot the residuals, use function ss3diags::SSrunstest.
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
