% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdiagsMCMC.R
\name{SSreadMCMC}
\alias{SSreadMCMC}
\title{Reads in Stock Synthesis MCMC file}
\usage{
SSreadMCMC(mcmcdir)
}
\arguments{
\item{mcmcdir}{file path for folder with the derived_posteriors.sso file}
}
\value{
Stock Synthesis mcmc output file
}
\description{
Imports the Stock Synthesis derived posteriors output file for MCMC data.
}
\author{
Henning Winker (JRC-EC) & Laurence Kell (Sea++)
}
\keyword{MCMC}
\keyword{utils}
