% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSretroComps.R
\name{SSretroComps}
\alias{SSretroComps}
\title{Retrieve Composition Data from Multiple Models}
\usage{
SSretroComps(retroModels)
}
\arguments{
\item{retroModels}{object list of replists from r4ss::SSgetoutput()}
}
\value{
list of observed and expected mean Length/age comps (c.f. ss3rep$cpue)
}
\description{
Wrapper to get observed and expected mean length/age from multiple Stock Synthesis models.
}
\author{
Henning Winker
}
\keyword{diags}
\keyword{retrocomps}
