% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSrmse.R
\name{SSrmse}
\alias{SSrmse}
\title{Function for calculating RMSE
used for `SSplotJABBAres()`}
\usage{
SSrmse(ss3rep, quants = c("cpue", "len", "age", "con"))
}
\arguments{
\item{ss3rep}{output from `SS_output()`}

\item{quants}{the dataset to calculate RMSE for. "cpue" for index of abundance, "len" for length comp, "age" for age composition, and "con" for conditional age-at-length.}
}
\value{
returns a list that includes the RMSE table output (by fleet and combined) and the dataframe of residuals which can be used for creating the `SSplotJABBAres()` plot
}
\description{
Function for calculating RMSE
used for `SSplotJABBAres()`
}
