% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSsettingsBratioF.R
\name{SSsettingsBratioF}
\alias{SSsettingsBratioF}
\title{SSsettingsBratioF}
\usage{
SSsettingsBratioF(ss3rep, status = c("Bratio", "F"), verbose = TRUE)
}
\arguments{
\item{ss3rep}{from r4ss::SSgetoutput()}

\item{status}{covarying stock status quantaties to extract from Hessian}

\item{verbose}{Report progress to R GUI?}
}
\value{
output list with Bratio and F settings
}
\description{
function to get Bratio and F settings
}
\author{
Henning Winker (JRC-EC)
}
