% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pac.hke}
\alias{pac.hke}
\title{2020 Pacific Hake Stock Synthesis model}
\format{
\code{r4ss} list objects
\describe{
 \item{ss3phke}{output from Stock Synthesis as read by r4ss::SS_output()}
 \item{retro.phk}{list of retrospective runs with r4ss::SS_doRetro() as read by r4ss::SSgetoutput()}
 \item{aspm.phk}{list of Stock Synthesis referece and aspm created with r4ss::SSsummarize()}
}
}
\source{
2020 Pacific hake/whiting Stock Synthesis assessment. Joint Technical Committee of the U.S. and Canada Pacific Hake/Whiting 
Agreement, National Marine Fisheries Service and Fisheries and Oceans Canada
}
\usage{
pac.hke
}
\description{
Stock Synthesis (V3.30.14) base-case run, list of retrospective runs,
 list of base-case and aspm runs
}
\author{
Grandin, C.J., K.F. Johnson, A.M. Edwards, and A.M. Berger. (2020). Status of the Pacific Hake (whiting) stock in U.S. and Canadian waters in 2020. Prepared by the Joint Technical Committee of the U.S. and Canada Pacific Hake/Whiting Agreement, National Marine Fisheries Service and Fisheries and Oceans Canada. 273 p.
}
