% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{sscol}
\alias{sscol}
\title{r4ss color generator}
\usage{
sscol(n, alpha = 1)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transluscency}
}
\value{
vector of color codes
}
\description{
r4ss color generator
}
