% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdeltaMVLN.R
\name{SSdeltaMVLN}
\alias{SSdeltaMVLN}
\title{Joint Distributions for Stock Status Ratios}
\usage{
SSdeltaMVLN(
  ss3rep,
  Fref = NULL,
  years = NULL,
  mc = 5000,
  weight = 1,
  run = "MVLN",
  plot = TRUE,
  addprj = FALSE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE
)
}
\arguments{
\item{ss3rep}{from r4ss::SS_output}

\item{Fref}{Choice of Fratio c("MSY","Btgt"), correponding to F_MSY and F_Btgt}

\item{years}{single year or vector of years for mvln}

\item{mc}{number of monte-carlo simulations}

\item{weight}{weighting option for model ensembles weight*mc}

\item{run}{qualifier for model run}

\item{plot}{option to show plot}

\item{addprj}{include forecast years}

\item{ymax}{ylim maximum}

\item{xmax}{xlim maximum}

\item{legendcex}{Allows to adjust legend cex}

\item{verbose}{Report progress to R GUI?}
}
\value{
output list of maximum likelihood estimates and the MVLN Monte-Carlo distributions of the Kobe values, and kobe plot
}
\description{
A function to generate joint distributions for stock status ratios (SSB/SSBref and F/Fref where ref can be MSY, SSB40, F40, etc.) using a Multivariate Log-Normal Distribution. The function produces a Kobe plot, maximum likelihood estimates and the MVLN Monte-Carlo distributions of the Kobe values which can be input into SSplotEnsemble().
}
\author{
Henning Winker (JRC-EC)
}
\keyword{diags}
\keyword{kobe}
\keyword{lognormal}
