% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SShcbias.R
\name{SShcbias}
\alias{SShcbias}
\title{Mohn's Rho and Forecast Bias}
\usage{
SShcbias(
  summaryoutput,
  quants = c("SSB", "F"),
  models = "all",
  endyrvec = "default",
  verbose = TRUE
)
}
\arguments{
\item{summaryoutput}{List created by r4ss::SSsummarize()}

\item{quants}{SSB or F quantity}

\item{models}{Optional subset of the models described in
r4ss function summaryoutput().  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{verbose}{Report progress to R GUI?}
}
\description{
This function produces the statistics from retrospective analysis (Mohn's rho and forecast bias). To visualize the retrospective forecasts from a Stock Synthesis model, use SSplotRetro().
}
\seealso{
[SSplotRetro()]
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
\keyword{diags}
\keyword{hindcasting}
