% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSmase.R
\name{SSmase}
\alias{SSmase}
\title{SSmase() computes MASE for one-step ahead hindcasting cross-validations of indices}
\usage{
SSmase(
  retroSummary,
  quants = c("cpue", "len", "age"),
  Season = "default",
  models = "all",
  endyrvec = "default",
  indexselect = NULL,
  MAE.base.adj = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{retroSummary}{List created by r4ss::SSsummarize()}

\item{quants}{data type c("cpue","len","age)}

\item{Season}{option to specify Season - Default uses first available, i.e. usual Seas = 1}

\item{models}{Optional subset of the models described in
r4ss function summaryoutput().  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{indexselect}{= Vector of fleet numbers for each model for which to compare}

\item{MAE.base.adj}{minimum MASE demoninator (naive predictions) for MASE.adj (default = 0.1)}

\item{verbose}{Report progress to R GUI?}

\item{indexfleets}{CHECK IF NEEDED or how to adjust indexfleets}
}
\value{
MASE and hcxval statistic
}
\description{
MASE for one-step ahead hindcasting cross-validations and computes MASE from prediction redisuals. 
MASE is calculated the average ratio of mean absolute error (MAE) of prediction residuals (MAE.PR) and Naive Predictions (MAE.base)
MASE.adj sets the MAE.base to a minimum MAE.base.adj (default=0.1)
MASE.adj allow passing (MASE<1) if MAE.PE < 0.1 and thus accurate, when obs show extremely little variation
}
\author{
Henning Winker (JRC-EC) and Laurence Kell (Sea++)
}
