% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSnew_run.R
\name{SSnewrun}
\alias{SSnewrun}
\title{SSnewrun()}
\usage{
SSnewrun(
  model,
  dat,
  ctl,
  ss.exe = "ss3.exe",
  newdir,
  overwrite = TRUE,
  pars = FALSE,
  par.file = "ss3.par"
)
}
\arguments{
\item{model}{model name folder to copy from}

\item{dat}{data file name, default "data.ss"}

\item{ctl}{control file name, default "control.ss"}

\item{ss.exe}{name of ss.exe file, default "control.exe"}

\item{newdir}{file path to copy the ss3 files to}

\item{overwrite}{if TRUE existingg files are overwritten}

\item{pars}{if TRUE the "ss3.par" is copied}

\item{par.file}{par file name, default ss3.par}
}
\description{
creates a subfolder with relevant ss3 file for retros and forecasts
}
\author{
Henning Winker
}
