% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotEnsemble.R
\name{SSplotEnsemble}
\alias{SSplotEnsemble}
\title{Model ensemble plots}
\usage{
SSplotEnsemble(
  kb,
  subplots = c("stock", "harvest", "SSB", "F", "Recr", "Catch"),
  models = "all",
  quantiles = list(c(0.025, 0.975)),
  ylabs = NULL,
  endyrvec = "default",
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = FALSE,
  use_pdf = FALSE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 5,
  tickEndYr = FALSE,
  xlim = NULL,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "l",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 12,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1, family = "sans"),
  verbose = TRUE,
  shadealpha = 0.3,
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  indexQlabel = TRUE,
  indexQdigits = 4,
  legendindex = NULL
)
}
\arguments{
\item{kb}{\code{kb} type output created by \code{SSdeltaMVLN()}}

\item{subplots}{vector to create subplots with these options:
\itemize{
\item \code{"stock"} Fish Population
\item \code{"harvest"} Harvest Rate
\item \code{"SSB"} Spawning Stock Biomass
\item \code{"F"} Fishing Mortality
\item \code{"Recr"} Recruitment
\item \code{"Catch"} Total Catch
}}

\item{models}{option to manually subset the models in \code{kb[["run"]]}}

\item{quantiles}{quantiles for uncertainty in plots. Input as a list,
default is the 95TH percentile: \code{list(c(0.025, 0.975))}}

\item{ylabs}{y-axis labels for quants}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By \code{"default"} it is set to the
ending year specified in each model.}

\item{plot}{DEPRECATED. By default, TRUE, Plots (and subplots) are drawn to the
plot device. The option to explicitly disable this option (FALSE), is not
implemented. This option flag will be defunct in a future version}

\item{print}{DEPRECATED, please use \code{print_plot}.}

\item{print_plot}{Flag to enable plot graphic device to print to PNG or PNG files.}

\item{png}{DEPRECATED. Please use \code{use_png}.}

\item{use_png}{Enables plots to be generated to PNG files. Defaults to print value}

\item{pdf}{DEPRECATED. Please use \code{use_pdf}.}

\item{use_pdf}{Enables plots to be generated to pdf file.}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{\link[base:logical]{Logical} flag: set TRUE or FALSE to switch to turn
on/off extra axis mark at final year in timeseries plots.}

\item{xlim}{Optional, years to use for x-axis. Default value NULL (or "default"),
uses all years available.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space.}

\item{xaxs}{Choice of xaxs parameter See \code{?par} for more info.}

\item{yaxs}{Choice of yaxs parameter. See \code{?par} for more info.}

\item{xylabs}{\link[base:logical]{Logical} flag: set TRUE or FALSE to include x- and
y-axis labels to the plot. Defaults to TRUE}

\item{type}{The type of plot to be drawn. For more details, see \code{\link[base:plot]{plot}}.}

\item{uncertainty}{Logical flag to enable plots with uncertainty intervals. Either a single
TRUE/FALSE value, or a vector of TRUE/FALSE values around SSB or F
estimated timeseries, or a set of integers corresponding to
the choice of models.}

\item{legend}{Flag to enable legend to plot. TRUE by default.}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector of two
numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See \code{help("legend")} for more info on the
string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have the legend
display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendcex}{Allows to adjust legend cex}

\item{legendsp}{Space between legend labels}

\item{pwidth}{Default width of plot printed to plot in units of \code{punits}}

\item{pheight}{Height of plot printed to plot in units of \code{punits }}

\item{punits}{Measurement units for \code{pwidth} and \code{pheight}. Default is \code{"in"}.
\itemize{
\item \code{"px"} (pixels)
\item \code{"in"} (inches)
\item \code{"cm"} (centimeters)
\item \code{"mm"} (millimeters)
}}

\item{res}{Resolution for plots printed to files.}

\item{ptsize}{Point size for plotted text in plots printed in files. See \code{help("png")}
for more details}

\item{cex.main}{Character expansion for plot titles.}

\item{plotdir}{Directory where output plot file will be written. By default,
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to output plot file name.
It will be separated from default name by an underscore.}

\item{par}{A numerical vector of the form c(bottom, left, top, right) which
gives the number of lines of margin to be specified on the four sides of
the plot,which is passed to \code{par()}. Entering \code{NULL} passes plot's default
\code{par()} values (which depends on whether the main title is included or not)}

\item{verbose}{Flag to print additional diagnostic messages to R console}

\item{shadealpha}{Transparency adjustment used to make uncertainty regions,
default is 0.3}

\item{new}{Deprecated. New plot windows are created by default (TRUE), and the
option to disable this, via FALSE, is unused.}

\item{add}{suppresses \code{par()} to create multiplot figs}

\item{mcmcVec}{Logical vector of TRUE/FALSE values (or single value) indicating
whether input values are from MCMC or to use normal distribution around
MLE.}

\item{indexQlabel}{TRUE/FALSE include labels for indices. Default is TRUE
(currently not used)}

\item{indexQdigits}{Number of significant digits for catchability in legend.
Default is 4}

\item{legendindex}{Allows to add legend for selected indices (plots)}
}
\description{
Plots model ensembles and forecasts with uncertainty represented by MVLN or
MCMC posteriors
}
\examples{
\dontrun{

mvln <- SSdeltaMVLN(simple, run = "Simple")
sspar(mfrow = c(3, 2), plot.cex = 0.7)
SSplotEnsemble(mvln[["kb"]], ylabs = mvln[["labels"]], add = T, verbose = F)
}
}
\author{
Mostly adopted from r4ss::SSplotComparisons by Taylor et al
}
\keyword{ensemble}
\keyword{ssplot}
