% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotGeneric.R
\name{SSplotGeneric}
\alias{SSplotGeneric}
\title{Generic SSplot function}
\usage{
SSplotGeneric(
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE
)
}
\arguments{
\item{pwidth}{Width of plot}

\item{pheight}{Height of plot}

\item{punits}{Measurement units for PNG file}

\item{res}{Resolution for PNG file}

\item{ptsize}{Point size for PNG file}

\item{cex.main}{Character expansion for plot titles}

\item{plotdir}{Directory where PNG or PDF files will be written. By default,
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{par}{list of graphics parameter values passed to \code{par()} function}

\item{verbose}{Flag to print additional diagnostic messages to R console}
}
\description{
Generalized version of ss3diags SSplots functions. Not intended for
production
}
