% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotHCxval.R
\name{SSplotHCxval}
\alias{SSplotHCxval}
\title{Hindcasting Cross-Validations of Multiple Models}
\usage{
SSplotHCxval(
  retroSummary,
  subplots = c("cpue", "len", "age"),
  Season = "default",
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  models = "all",
  endyrvec = "default",
  xmin = NULL,
  indexselect = NULL,
  MAE.base.adj = 0.1,
  show.mase.adj = TRUE,
  indexUncertainty = TRUE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  xlim = "default",
  ylimAdj = 1.15,
  ylim = NULL,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = grey(0.5, 0.4),
  shadecol2 = grey(0.5, 0.4),
  shadealpha = 0.3,
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  indexQlabel = TRUE,
  indexQdigits = 4,
  indexfleets = 1,
  plot = TRUE,
  shadecol1 = grey(0.5, 0.4)
)
}
\arguments{
\item{retroSummary}{Stock synthesis summary list created by \code{\link[r4ss:SSsummarize]{SSsummarize}}
or a summary list from age/length retrospective runs created by \link[=SSretroComps]{'SSretroComps'}}

\item{subplots}{optional use of c("cpue","len","age"), yet to be tested for age.}

\item{Season}{option to specify Season as an integer of value 1-4 - Default uses first available, i.e. usual Seas = 1}

\item{print}{Deprecated. Please use \code{print_plot}}

\item{print_plot}{Option to output plot as output image file.}

\item{png}{Deprecated. Please use \code{use_png}.}

\item{use_png}{Option to use the \code{\link[grDevices:png]{png}} graphical device to draw plot}

\item{pdf}{Deprecated. Please use 'use_pdf'.}

\item{use_pdf}{Option to use the \code{\link[grDevices:pdf]{pdf}} graphical device to generate pdf plots. (Note: not tested  TRUE)}

\item{models}{Optional subset of the r4ss summary output object models described in r4ss
function \code{\link[r4ss:SSsummarize]{SSsummarize}}().  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{xmin}{optional number first year shown in plot (if available)}

\item{indexselect}{= Vector of fleet numbers for each model for which to compare}

\item{MAE.base.adj}{minimum MASE denominator (naive predictions) for MASE.adj (default = 0.1)}

\item{show.mase.adj}{if TRUE it show mase.adj in () in plot}

\item{indexUncertainty}{Show fixed uncertainty intervals on index (not estimated)}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional \code{\link[graphics:par]{par}} vector of plot character values}

\item{lty}{Optional \code{\link[graphics:par]{par}} vector of line types}

\item{lwd}{Optional \code{\link[graphics:par]{par}} vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis marks at final year in timeseries plots.}

\item{xlim}{Optional, values for x-axis range of years to display on plot. Default = "default" displays all years of available data. (currently not in use)}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space.}

\item{ylim}{will over-write ylimAdj if specified}

\item{xaxs}{Choice of \code{\link[graphics:par]{par}} xaxs parameter (see \code{\link[graphics:par]{par}} for more info)}

\item{yaxs}{Choice of \code{\link[graphics:par]{par}} yaxs parameter (see \code{\link[graphics:par]{par}} for more info)}

\item{xylabs}{Option to include x- and y-axis labels. Defaults to TRUE}

\item{type}{Type parameter passed to points. If type is \code{o}, it overplots points on
top of lines. if type input is \code{l}, then turn off points on top of
lines in legend. For more detail see \code{\link[graphics:plot.default]{plot.default}}}

\item{uncertainty}{\link[base:logical]{Logical vector} (TRUE/FALSE) to include uncertainty intervals
around SSB or F estimated time series. Defaults to TRUE.}

\item{legend}{Option to add a legend. TRUE by default.}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like \code{topleft} or a vector of two
numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See \code{\link[graphics:legend]{?legend}} for more info on the
string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have the legend
display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend}

\item{legendcex}{Allows to adjust \code{\link[graphics:legend]{legend}} cex}

\item{legendsp}{Space between legend labels}

\item{legendindex}{Allows to add legend for selected indices (plots)}

\item{pwidth}{Width of the plot}

\item{pheight}{Height of the plot}

\item{punits}{Measurement units for output PNG image file}

\item{res}{Resolution for output PNG image file}

\item{ptsize}{Point size for output PNG image file}

\item{cex.main}{Character Expansion for plot titles. See \link[graphics:par]{'par'} for more info.}

\item{plotdir}{Directory where output image files will be written. By default,
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to output image file name. It will be separated from the
default name by an underscore.}

\item{par}{list of graphics parameter values passed to \link[graphics:par]{'par'} function.
Used to adjust the plot margins.}

\item{verbose}{Report progress to R GUI?}

\item{shadecol}{uncertainty shading of hcxval horizon}

\item{shadecol2}{color for uncertainty in early years not affected by hindcast}

\item{shadealpha}{Transparancy adjustment used to make default shadecol. (currently not used)}

\item{new}{Deprecated. New plot windows are created by default (\code{TRUE}), and the
option to disable this, via FALSE, is unused.}

\item{add}{suppresses \code{\link[graphics:par]{par}} to create multiplot figs}

\item{mcmcVec}{UNTESTED. Vector of TRUE/FALSE values (or single value) indicating mcmc
values are used.}

\item{indexQlabel}{Logical; TRUE/FALSE. Option to include labels for indices. Default is TRUE (currently not used)}

\item{indexQdigits}{Number of significant digits for catchability in legend. Default is 4}

\item{indexfleets}{CHECK IF NEEDED or how to adjust indexfleets}

\item{plot}{Deprecated. Plots (and subplots) are drawn to the active plot device
by default (TRUE), and the option to disable this, via FALSE, is unused.}

\item{shadecol1}{uncertainty shading of early years not affected by hindcast (currently not used)}
}
\description{
Plots one-step ahead hindcasting cross-validations and computes MASE from prediction residuals.
MASE is calculated as the average ratio of mean absolute error (MAE) of prediction residuals (MAE.PR) and naive predictions (MAE.base)
MASE.adj sets the MAE.base to a minimum MAE.base.adj (default=0.1)
MASE.adj allow passing (MASE<1) if MAE.PE < 0.1 and thus accurate if obs show very little annual variation
}
\author{
Henning Winker (JRC-EC) and Laurence Kell (Sea++)
}
\keyword{hindcasting}
\keyword{ssplot}
