% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotHCxval.R
\name{SSplotHCxval}
\alias{SSplotHCxval}
\title{Hindcasting Cross-Validations of Multiple Models}
\usage{
SSplotHCxval(
  retroSummary,
  subplots = c("cpue", "len", "age"),
  Season = "default",
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  models = "all",
  endyrvec = "default",
  xmin = NULL,
  indexselect = NULL,
  MAE.base.adj = 0.1,
  show.mase.adj = TRUE,
  indexUncertainty = TRUE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  xlim = "default",
  ylimAdj = 1.15,
  ylim = NULL,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.9,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 12,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1, family = "sans"),
  verbose = TRUE,
  shadecol = grey(0.5, 0.4),
  shadecol2 = grey(0.5, 0.4),
  shadealpha = 0.3,
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  indexQlabel = TRUE,
  indexQdigits = 4,
  indexfleets = 1,
  plot = TRUE,
  shadecol1 = grey(0.5, 0.4)
)
}
\arguments{
\item{retroSummary}{List created by \code{\link[r4ss:SSsummarize]{r4ss::SSsummarize()}} or
\code{\link[=SSretroComps]{SSretroComps()}}}

\item{subplots}{optional use of the following:
\itemize{
\item \code{"cpue"} Index of abundance data
\item \code{"len"} Length-composition data
\item \code{"age"} Age-composition data
}}

\item{Season}{option to specify Season as an integer of value 1-4. Default
uses first available, i.e. usual Seas = 1}

\item{print}{DEPRECATED, please use \code{print_plot}.}

\item{print_plot}{Flag to enable plot graphic device to print to PNG or PNG files.}

\item{png}{DEPRECATED. Please use \code{use_png}.}

\item{use_png}{Enables plots to be generated to PNG files. Defaults to print value}

\item{pdf}{DEPRECATED. Please use \code{use_pdf}.}

\item{use_pdf}{Enables plots to be generated to pdf file.}

\item{models}{Optional subset of the models of \code{summaryoutput} (or a similar
field with a different name): a list created by the function
\link[r4ss:SSsummarize]{r4ss::SSsummarize}. Either \code{"all"} or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{xmin}{optional number first year shown in plot (if available)}

\item{indexselect}{Vector of fleet numbers for each model for which to
compare}

\item{MAE.base.adj}{minimum \code{MASE} denominator (naive predictions) for
\code{MASE.adj} (default = 0.1)}

\item{show.mase.adj}{if TRUE, it show \code{mase.adj} in plot}

\item{indexUncertainty}{Show fixed uncertainty intervals on index
(not estimated)}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{\link[base:logical]{Logical} flag: set TRUE or FALSE to switch to turn
on/off extra axis mark at final year in timeseries plots.}

\item{xlim}{Optional, years to use for x-axis. Default value NULL (or "default"),
uses all years available.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space.}

\item{ylim}{will over-write \code{ylimAdj} if specified}

\item{xaxs}{Choice of xaxs parameter See \code{?par} for more info.}

\item{yaxs}{Choice of yaxs parameter. See \code{?par} for more info.}

\item{xylabs}{\link[base:logical]{Logical} flag: set TRUE or FALSE to include x- and
y-axis labels to the plot. Defaults to TRUE}

\item{type}{The type of plot to be drawn. For more details, see \code{\link[base:plot]{plot}}.}

\item{uncertainty}{Logical flag to enable plots with uncertainty intervals. Either a single
TRUE/FALSE value, or a vector of TRUE/FALSE values around SSB or F
estimated timeseries, or a set of integers corresponding to
the choice of models.}

\item{legend}{Flag to enable legend to plot. TRUE by default.}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector of two
numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See \code{help("legend")} for more info on the
string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have the legend
display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendcex}{Allows to adjust legend cex}

\item{legendsp}{Space between legend labels}

\item{legendindex}{Allows to add legend for selected indices (plots)}

\item{pwidth}{Default width of plot printed to plot in units of \code{punits}}

\item{pheight}{Height of plot printed to plot in units of \code{punits }}

\item{punits}{Measurement units for \code{pwidth} and \code{pheight}. Default is \code{"in"}.
\itemize{
\item \code{"px"} (pixels)
\item \code{"in"} (inches)
\item \code{"cm"} (centimeters)
\item \code{"mm"} (millimeters)
}}

\item{res}{Resolution for plots printed to files.}

\item{ptsize}{Point size for plotted text in plots printed in files. See \code{help("png")}
for more details}

\item{cex.main}{Character expansion for plot titles.}

\item{plotdir}{Directory where output plot file will be written. By default,
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to output plot file name.
It will be separated from default name by an underscore.}

\item{par}{A numerical vector of the form c(bottom, left, top, right) which
gives the number of lines of margin to be specified on the four sides of
the plot,which is passed to \code{par()}. Entering \code{NULL} passes plot's default
\code{par()} values (which depends on whether the main title is included or not)}

\item{verbose}{Flag to print additional diagnostic messages to R console}

\item{shadecol}{uncertainty shading of hcxval horizon}

\item{shadecol2}{color for uncertainty in early years not affected by
hindcast}

\item{shadealpha}{Transparency adjustment used to make default \code{shadecol}.
(currently not used)}

\item{new}{Deprecated. New plot windows are created by default (TRUE), and the
option to disable this, via FALSE, is unused.}

\item{add}{suppresses \code{par()} to create multiplot figs}

\item{mcmcVec}{Logical vector of TRUE/FALSE values (or single value) indicating
whether input values are from MCMC or to use normal distribution around
MLE.}

\item{indexQlabel}{\link[base:logical]{Logical} flag to add catchability to
legend in plot of index fits.}

\item{indexQdigits}{Number of significant digits for catchability in legend}

\item{indexfleets}{Fleet numbers for each model to compare
indices of abundance. Can take different forms:
\itemize{
\item integer: (default) create a single comparison plot for the chosen
index
\item NULL: create a separate plot for each index as long as the fleet
numbering is the same across all models.
\item vector of length equal to number of models: a single fleet number
for each model to be compared in a single plot
\item list: list of fleet numbers associated with indices within each
model to be compared, where the list elements are each a vector of the
same length but the names of the list elements don't matter and can be
absent.
}}

\item{plot}{DEPRECATED. By default, TRUE, Plots (and subplots) are drawn to the
plot device. The option to explicitly disable this option (FALSE), is not
implemented. This option flag will be defunct in a future version}

\item{shadecol1}{uncertainty shading of early years not affected by
hindcast (currently not used)}
}
\description{
Plots one-step ahead hindcasting cross-validations and computes MASE from
prediction residuals.\code{MASE} is calculated as the average ratio of mean
absolute error (\code{MAE}) of prediction residuals (\code{MAE.PR}) and naive
predictions (\code{MAE.base}). \code{MASE.adj} sets the \code{MAE.base} to a minimum
\code{MAE.base.adj} (\code{default=0.1}). \code{MASE.adj} allow passing (\code{MASE<1}) if
\code{MAE.PE < 0.1} and thus accurate if obs show very little annual variation
}
\author{
Henning Winker (JRC-EC) and Laurence Kell (Sea++)
}
\keyword{hindcasting}
\keyword{ssplot}
