% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRetro.R
\name{SSplotRetro}
\alias{SSplotRetro}
\title{Retrospective-Forecast with One-Step Ahead Hindcasting}
\usage{
SSplotRetro(
  summaryoutput,
  subplots = c("SSB", "F"),
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  models = "all",
  endyrvec = "default",
  xlim = NULL,
  xmin = NULL,
  labels = NULL,
  ylim = NULL,
  forecast = TRUE,
  forecastrho = TRUE,
  showrho = TRUE,
  col = NULL,
  pch = NULL,
  lty = 1,
  lwd = 2,
  tickEndYr = TRUE,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 0.7,
  legendindex = NULL,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  shadecol = grey(0.4, 0.6),
  new = TRUE,
  add = FALSE,
  mcmcVec = FALSE,
  shadecol1 = grey(0.5, 0.4),
  indexQlabel = TRUE,
  indexQdigits = 4,
  shadealpha = 0.3
)
}
\arguments{
\item{summaryoutput}{List created by r4ss::SSsummarize()}

\item{subplots}{Optional vector of subplots to be created, "SSB", "Bratio", "Fvalue", "Recruits", "Index"}

\item{plot}{Deprecated. Plots (and subplots) are drawn to the active plot device
by default (TRUE), and the option to disable this, via FALSE, is unused.}

\item{print}{Deprecated. Please use \code{print_plot}}

\item{print_plot}{Option to output plot as output image file.}

\item{png}{Deprecated. Please use \code{use_png}.}

\item{use_png}{Option to use the \code{\link[grDevices:png]{png}} graphical device to draw plot}

\item{pdf}{Deprecated. Please use 'use_pdf'.}

\item{use_pdf}{Option to use the \code{\link[grDevices:pdf]{pdf}} graphical device to generate pdf plots. (Note: not tested  TRUE)}

\item{models}{Optional subset of the r4ss summary output object models described in r4ss
function \code{\link[r4ss:SSsummarize]{SSsummarize}}().  Either "all" or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{xlim}{optional xlim, which overwrites xmin}

\item{xmin}{optional minimum year shown in plot (default first yr)}

\item{labels}{yaxis lable for biomass (bony fish and sharks)}

\item{ylim}{option to specify ylim range}

\item{forecast}{if true one-step ahead forecasts are shown in plot}

\item{forecastrho}{if true one-step ahead forecast rho value is denoted in plot}

\item{showrho}{TRUE/FALSE include Mohn's rho value? Defaults to TRUE}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional \code{\link[graphics:par]{par}} vector of plot character values}

\item{lty}{Optional \code{\link[graphics:par]{par}} vector of line types}

\item{lwd}{Optional \code{\link[graphics:par]{par}} vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis marks at final}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space.}

\item{xaxs}{Choice of \code{\link[graphics:par]{par}} xaxs parameter (see \code{\link[graphics:par]{par}} for more info)}

\item{yaxs}{Choice of \code{\link[graphics:par]{par}} yaxs parameter (see \code{\link[graphics:par]{par}} for more info)}

\item{xylabs}{TRUE or FALSE, include x- and y-axis labels. Defaults to TRUE}

\item{type}{Type parameter passed to points. If type is \code{o}, it overplots points on
top of lines. if type input is \code{l}, then turn off points on top of
lines in legend. For more detail see \code{\link[graphics:plot.default]{plot.default}}}

\item{uncertainty}{\link[base:logical]{Logical vector} (TRUE/FALSE) to include uncertainty intervals
around SSB or F estimated time series. Defaults to TRUE.}

\item{legend}{Option to add a legend. TRUE by default.}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like \code{topleft} or a vector of two
numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See \code{\link[graphics:legend]{?legend}} for more info on the
string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have the legend
display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend}

\item{legendcex}{Allows to adjust \code{\link[graphics:legend]{legend}} cex}

\item{legendsp}{Space between legend labels}

\item{legendindex}{Allows to add legend for selected indices (plots)}

\item{pwidth}{Width of the plot}

\item{pheight}{Height of the plot}

\item{punits}{Measurement units for output PNG image file}

\item{res}{Resolution for output PNG image file}

\item{ptsize}{Point size for output PNG image file}

\item{cex.main}{Character Expansion for plot titles. See \link[graphics:par]{'par'} for more info.}

\item{plotdir}{Directory where output image files will be written. By default,
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to output image file name. It will be separated from the
default name by an underscore.}

\item{par}{list of graphics parameter values passed to \link[graphics:par]{'par'} function.
Used to adjust the plot margins.}

\item{verbose}{Option to report progress to R console. Default is TRUE.}

\item{shadecol}{uncertainty shading of hcxval horizon}

\item{new}{Deprecated. New plot windows are created by default (\code{TRUE}), and the
option to disable this, via FALSE, is unused.}

\item{add}{suppresses \code{\link[graphics:par]{par}} to create multiplot figs}

\item{mcmcVec}{UNTESTED. Vector of TRUE/FALSE values (or single value) indicating mcmc
values are used.}

\item{shadecol1}{uncertainty shading of early years not affected by hindcast}

\item{indexQlabel}{Logical; TRUE/FALSE. Option to include labels for indices. Default is TRUE (currently not used)}

\item{indexQdigits}{Number of significant digits for catchability in legend. Default is 4}

\item{shadealpha}{set the transparency level (alpha) of the area of uncertainty. Defalut to 0.3 (currently not used)}
}
\description{
Plots retrospective pattern, including (optional) one-step ahead forecast and computes Mohn's Rho
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
\keyword{retro}
\keyword{ssplot}
