% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRetro.R
\name{SSplotRetro}
\alias{SSplotRetro}
\title{Retrospective-Forecast with One-Step Ahead Hindcasting}
\usage{
SSplotRetro(
  summaryoutput,
  subplots = c("SSB", "F"),
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  models = "all",
  endyrvec = "default",
  xlim = NULL,
  xmin = NULL,
  labels = NULL,
  ylim = NULL,
  forecast = TRUE,
  forecastrho = TRUE,
  showrho = TRUE,
  col = NULL,
  pch = NA,
  lwd = 5,
  tickEndYr = TRUE,
  ylimAdj = 1.05,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  uncertainty = TRUE,
  legend = TRUE,
  legendlabels = "default",
  legendloc = "topright",
  legendorder = "default",
  legendncol = 1,
  legendcex = 1,
  legendsp = 1,
  legendindex = NULL,
  pt.cex = 3,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 12,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1, family = "sans"),
  verbose = TRUE,
  shadecol = "#404040",
  new = TRUE,
  add = TRUE,
  mcmcVec = FALSE,
  shadecol1 = "#c8d0d9",
  indexQlabel = TRUE,
  indexQdigits = 4,
  shadealpha = 0.3
)
}
\arguments{
\item{summaryoutput}{List created by \code{\link[r4ss:SSsummarize]{r4ss::SSsummarize()}}}

\item{subplots}{Optional vector of subplots to be created:
\itemize{
\item \code{"SSB"} Spawning Stock Biomass
\item \code{"F"} Fishing Mortality
}}

\item{plot}{DEPRECATED. By default, TRUE, Plots (and subplots) are drawn to the
plot device. The option to explicitly disable this option (FALSE), is not
implemented. This option flag will be defunct in a future version}

\item{print}{DEPRECATED, please use \code{print_plot}.}

\item{print_plot}{Flag to enable plot graphic device to print to PNG or PNG files.}

\item{png}{DEPRECATED. Please use \code{use_png}.}

\item{use_png}{Enables plots to be generated to PNG files. Defaults to print value}

\item{pdf}{DEPRECATED. Please use \code{use_pdf}.}

\item{use_pdf}{Enables plots to be generated to pdf file.}

\item{models}{Optional subset of the models of \code{summaryoutput} (or a similar
field with a different name): a list created by the function
\link[r4ss:SSsummarize]{r4ss::SSsummarize}. Either \code{"all"} or a vector of numbers indicating
columns in summary tables.}

\item{endyrvec}{Optional single year or vector of years representing the
final year of values to show for each model. By default it is set to the
ending year specified in each model.}

\item{xlim}{Optional, years to use for x-axis. Default value NULL (or "default"),
uses all years available.}

\item{xmin}{optional minimum year shown in plot (default first yr)}

\item{labels}{\code{yaxis} label for biomass (bony fish and sharks)}

\item{ylim}{option to specify \code{ylim} range}

\item{forecast}{\link[base:logical]{Logical}. If TRUE, one-step ahead forecasts
are shown in plot}

\item{forecastrho}{\link[base:logical]{Logical}. If TRUE, one-step ahead
forecast rho value is denoted in plot}

\item{showrho}{\link[base:logical]{Logical} flag to include Mohn's rho value.
Defaults to TRUE}

\item{col}{Optional vector of colors to be used for lines. Input NULL}

\item{pch}{Optional vector of plot character values}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{\link[base:logical]{Logical} flag: set TRUE or FALSE to switch to turn
on/off extra axis mark at final year in timeseries plots.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space.}

\item{xaxs}{Choice of xaxs parameter See \code{?par} for more info.}

\item{yaxs}{Choice of yaxs parameter. See \code{?par} for more info.}

\item{xylabs}{\link[base:logical]{Logical} flag: set TRUE or FALSE to include x- and
y-axis labels to the plot. Defaults to TRUE}

\item{type}{The type of plot to be drawn. For more details, see \code{\link[base:plot]{plot}}.}

\item{uncertainty}{Logical flag to enable plots with uncertainty intervals. Either a single
TRUE/FALSE value, or a vector of TRUE/FALSE values around SSB or F
estimated timeseries, or a set of integers corresponding to
the choice of models.}

\item{legend}{Flag to enable legend to plot. TRUE by default.}

\item{legendlabels}{Optional vector of labels to include in legend.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector of two
numeric values representing the fraction of the maximum in the x and y
dimensions, respectively. See \code{help("legend")} for more info on the
string options.}

\item{legendorder}{Optional vector of model numbers that can be used to have the legend
display the model names in an order that is different than that
which is represented in the summary input object.}

\item{legendncol}{Number of columns for the legend.}

\item{legendcex}{Allows to adjust legend cex}

\item{legendsp}{Space between legend labels}

\item{legendindex}{Allows to add legend for selected indices (plots)}

\item{pwidth}{Default width of plot printed to plot in units of \code{punits}}

\item{pheight}{Height of plot printed to plot in units of \code{punits }}

\item{punits}{Measurement units for \code{pwidth} and \code{pheight}. Default is \code{"in"}.
\itemize{
\item \code{"px"} (pixels)
\item \code{"in"} (inches)
\item \code{"cm"} (centimeters)
\item \code{"mm"} (millimeters)
}}

\item{res}{Resolution for plots printed to files.}

\item{ptsize}{Point size for plotted text in plots printed in files. See \code{help("png")}
for more details}

\item{cex.main}{Character expansion for plot titles.}

\item{plotdir}{Directory where output plot file will be written. By default,
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to output plot file name.
It will be separated from default name by an underscore.}

\item{par}{A numerical vector of the form c(bottom, left, top, right) which
gives the number of lines of margin to be specified on the four sides of
the plot,which is passed to \code{par()}. Entering \code{NULL} passes plot's default
\code{par()} values (which depends on whether the main title is included or not)}

\item{verbose}{Flag to print additional diagnostic messages to R console}

\item{shadecol}{uncertainty shading of hcxval horizon}

\item{new}{Deprecated. New plot windows are created by default (TRUE), and the
option to disable this, via FALSE, is unused.}

\item{add}{suppresses \code{par()} to create multiplot figs}

\item{mcmcVec}{Logical vector of TRUE/FALSE values (or single value) indicating
whether input values are from MCMC or to use normal distribution around
MLE.}

\item{shadecol1}{uncertainty shading of early years not affected by hindcast}

\item{indexQlabel}{\link[base:logical]{Logical}. If TRUE, add catchability to
legend in plot of index fits (currently not used)}

\item{indexQdigits}{Number of significant digits for catchability in legend}

\item{shadealpha}{set the transparency level (alpha) of the area of
uncertainty. Defaults to 0.3 (currently not used)}
}
\description{
Plots retrospective pattern, including (optional) one-step ahead forecast
and computes Mohn's Rho
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
\keyword{retro}
\keyword{ssplot}
