% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRunstest.R
\name{SSplotRunstest}
\alias{SSplotRunstest}
\title{plot function for runs test plot}
\usage{
SSplotRunstest(
  ss3rep = ss3sma,
  mixing = "less",
  subplots = c("cpue", "len", "age")[1],
  plot = TRUE,
  print = FALSE,
  png = print,
  pdf = FALSE,
  indexselect = NULL,
  miny = 1,
  pch = 21,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = "default",
  ylimAdj = 1.4,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  legend = TRUE,
  legendloc = "top",
  legendcex = 1,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  new = TRUE,
  add = FALSE
)
}
\arguments{
\item{ss3rep}{from r4ss::SSgetoutput()$replist1}

\item{mixing}{c("less","greater","two.sided"). Default less is checking for postive autocorrelation only}

\item{subplots}{optional use of c("cpue","len","age"), yet to be tested for age.}

\item{plot}{plot to active plot device?}

\item{print}{print to PNG files?}

\item{pdf}{not tested for TRUE}

\item{indexselect}{Vector of fleet numbers for each model for which to compare}

\item{miny}{minimum abs values of ylim}

\item{pch}{Optional vector of plot character values}

\item{lty}{Optional vector of line types}

\item{lwd}{Optional vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis mark at final
year in timeseries plots.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space}

\item{xaxs}{Choice of xaxs parameter (see ?par for more info)}

\item{yaxs}{Choice of yaxs parameter (see ?par for more info)}

\item{type}{Type parameter passed to points (default 'o' overplots points on
top of lines)}

\item{legend}{Add a legend?}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector}

\item{pwidth}{Width of plot}

\item{pheight}{Height of plot}

\item{punits}{Units for PNG file}

\item{res}{Resolution for PNG file}

\item{ptsize}{Point size for PNG file}

\item{cex.main}{Character expansion for plot titles}

\item{plotdir}{Directory where PNG or PDF files will be written. By default
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to PNG or PDF file names.
It will be separated from default name by an underscore.}

\item{par}{list of graphics parameter values passed to par() function}

\item{verbose}{Report progress to R GUI?}

\item{new}{Create new empty plot window}

\item{add}{surpresses par() to create multiplot figs}

\item{legendcex=1}{Allows to adjust legend cex}
}
\value{
Runs Test p-values and sig3 limits
}
\description{
Residual diagnostics with runs test p-value and 3xsigma limits for Indices and meanL
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
