% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRunstest.R
\name{SSplotRunstest}
\alias{SSplotRunstest}
\title{Residual Diagnostics}
\usage{
SSplotRunstest(
  ss3rep = ss3diags::simple,
  mixing = "less",
  subplots = c("cpue", "len", "age", "size", "con")[1],
  plot = TRUE,
  print = deprecated(),
  print_plot = FALSE,
  png = deprecated(),
  use_png = print_plot,
  pdf = deprecated(),
  use_pdf = FALSE,
  indexselect = NULL,
  miny = 1,
  pch = 21,
  lty = 1,
  lwd = 2,
  tickEndYr = FALSE,
  xlim = "default",
  ylim = "default",
  ylimAdj = 1.4,
  xaxs = "i",
  yaxs = "i",
  xylabs = TRUE,
  type = "o",
  legend = TRUE,
  legendloc = "top",
  legendcex = 1,
  pwidth = 6.5,
  pheight = 5,
  punits = "in",
  res = 300,
  ptsize = 10,
  cex.main = 1,
  plotdir = NULL,
  filenameprefix = "",
  par = list(mar = c(5, 4, 1, 1) + 0.1),
  verbose = TRUE,
  new = TRUE,
  add = FALSE
)
}
\arguments{
\item{ss3rep}{Stock Synthesis output as read by r4SS function SS_output}

\item{mixing}{c("less","greater","two.sided"). Default less is checking for positive autocorrelation only}

\item{subplots}{optional 'cpue' for index data, 'len' for length composition data, 'size' for
generalized size composition data, 'age' for age composition data,
or 'con' for conditional age at length data}

\item{plot}{Deprecated. Plots (and subplots) are drawn to the active plot device
by default (TRUE), and the option to disable this, via FALSE, is unused.}

\item{print}{Deprecated. Please use \code{print_plot}}

\item{print_plot}{Option to output plot as output image file.}

\item{png}{Deprecated. Please use \code{use_png}.}

\item{use_png}{Option to use the \code{\link[grDevices:png]{png}} graphical device to draw plot}

\item{pdf}{Deprecated. Please use 'use_pdf'.}

\item{use_pdf}{Option to use the \code{\link[grDevices:pdf]{pdf}} graphical device to generate pdf plots. (Note: not tested  TRUE)}

\item{indexselect}{Vector of fleet numbers for each model for which to compare}

\item{miny}{the absolute value of the min and max value for ylim, default is 1}

\item{pch}{Optional \code{\link[graphics:par]{par}} vector of plot character values}

\item{lty}{Optional \code{\link[graphics:par]{par}} vector of line types}

\item{lwd}{Optional \code{\link[graphics:par]{par}} vector of line widths}

\item{tickEndYr}{TRUE/FALSE switch to turn on/off extra axis marks at final year in timeseries plots.}

\item{xlim}{Optional, values for x-axis range of years to display on plot. Default = "default" displays all years of available data.}

\item{ylim}{Optional, values for y-axis range to display on plot. Default = "default" will range from -1 to 1.}

\item{ylimAdj}{Multiplier for ylim parameter. Allows additional white space.}

\item{xaxs}{Choice of \code{\link[graphics:par]{par}} xaxs parameter (see \code{\link[graphics:par]{par}} for more info)}

\item{yaxs}{Choice of \code{\link[graphics:par]{par}} yaxs parameter (see \code{\link[graphics:par]{par}} for more info)}

\item{xylabs}{Option to include x- and y-axis labels. Defaults to TRUE}

\item{type}{Type parameter passed to points. If type is \code{o}, it overplots points on
top of lines. if type input is \code{l}, then turn off points on top of
lines in legend. For more detail see \code{\link[graphics:plot.default]{plot.default}}}

\item{legend}{Option to add a legend. TRUE by default.}

\item{legendloc}{Location of legend. Either a string like "topleft" or a vector}

\item{legendcex}{Allows to adjust legend cex}

\item{pwidth}{Width of the plot}

\item{pheight}{Height of the plot}

\item{punits}{Measurement units for output PNG image file}

\item{res}{Resolution for output PNG image file}

\item{ptsize}{Point size for output PNG image file}

\item{cex.main}{Character Expansion for plot titles. See \link[graphics:par]{'par'} for more info.}

\item{plotdir}{Directory where output image files will be written. By default,
it will be the directory where the model was run.}

\item{filenameprefix}{Additional text to append to output image file name. It will be separated from the
default name by an underscore.}

\item{par}{list of graphics parameter values passed to \link[graphics:par]{'par'} function.
Used to adjust the plot margins.}

\item{verbose}{Option to report progress to R console. Default is TRUE.}

\item{new}{Deprecated. New plot windows are created by default (\code{TRUE}), and the
option to disable this, via FALSE, is unused.}

\item{add}{suppresses \code{\link[graphics:par]{par}} to create multiplot figs}
}
\value{
a dataframe with runs test p-value, if the test has passed or failed, 3x sigma high and low limits, and the type of data used. Rows are for each fleet. Note, runs test passed if p-value > 0.05 (residuals are random) and failed if p-value < 0.5 (residuals are not random)
}
\description{
Function for residual diagnostics. Plots residuals and 3x sigma limits for indices or mean age or length and outputs a runs test table. Note, if you do not want to plot the residuals, use function ss3diags::SSrunstest.
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
\keyword{runsTest}
\keyword{ssplot}
