% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdiagsMCMC_test.R
\name{SSreadMCMC}
\alias{SSreadMCMC}
\title{SSreadMCMC()}
\usage{
SSreadMCMC(mcmcdir)
}
\arguments{
\item{mcmcdir}{file path for folder with the derived_posteriors.sso file}
}
\value{
Stock Synthesis mcmc output file
}
\description{
function to read mcmc file outputs
}
\author{
Henning Winker (JRC-EC) & Laurence Kell (Sea++)
}
