% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdiagsMCMC.R
\name{SSreadMCMC}
\alias{SSreadMCMC}
\title{Read SS MCMC}
\usage{
SSreadMCMC(mcmcdir)
}
\arguments{
\item{mcmcdir}{file path for folder with the derived_posteriors.sso file}
}
\value{
Stock Synthesis mcmc output file
}
\description{
A function to read mcmc file outputs from a Stock Synthesis model
}
\author{
Henning Winker (JRC-EC) & Laurence Kell (Sea++)
}
\keyword{MCMC}
\keyword{utils}
