% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSrmse.R
\name{SSrmse}
\alias{SSrmse}
\title{Function for calculating RMSE
used for \code{SSplotJABBAres()}}
\usage{
SSrmse(ss3rep, quants, seas = NULL, indexselect = NULL)
}
\arguments{
\item{ss3rep}{output from \code{SS_output()}}

\item{quants}{the dataset to calculate RMSE for. "cpue" for index of abundance, "len" for length comp, "age" for age composition, "size" for general size composition, and "con" for conditional age-at-length.}

\item{seas}{string indicating how to treat data from multiple seasons
'comb' - combine seasonal data for each year and plot against Yr
'sep' - treat season separately, plotting against Yr.S.
If is.null(seas), it is assumed that there is only one season and option 'comb' is used.}

\item{indexselect}{Vector of fleet numbers for each model for which to compare}
}
\value{
returns a list that includes the RMSE table output (by fleet and combined) and the dataframe of residuals which can be used for creating the \code{SSplotJABBAres()} plot
}
\description{
Function for calculating RMSE
used for \code{SSplotJABBAres()}
}
\keyword{rmsetable}
