% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSplotRunstest.R
\name{SSrunstest}
\alias{SSrunstest}
\title{runs test}
\usage{
SSrunstest(
  ss3rep = ss3diags::ss3sma,
  mixing = "less",
  quants = c("cpue", "len", "age")[1],
  indexselect = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ss3rep}{from r4ss::SSgetoutput()$replist1}

\item{mixing}{c("less","greater","two.sided"). Default less is checking for postive autocorrelation only}

\item{quants}{optional use of c("cpue","len","age"), yet to be tested for age.}

\item{indexselect}{Vector of fleet numbers for each model for which to compare}

\item{verbose}{Report progress to R GUI?}
}
\value{
Runs Test p-values and sig3 limits
}
\description{
Residual diagnostics with runs test p-value and 3xsigma limits for Indices, mean length and mean age
}
\author{
Henning Winker (JRC-EC) and Laurance Kell (Sea++)
}
\keyword{diags}
\keyword{runsTest}
